+++ Modules/Intent.Modules.Rdbms.Importer/release-notes.md
@@ -3,6 +3,7 @@
 - Improvement: Introducing `Attribute Name Convention` for how Column names translates into Entity attribute names.
 - Fixed: The `Filter Type` will now be remembered when the next Import has to happen.
 - Fixed: Incorrectly looked for duplicate associations.
+- Fixed: Stored procedure `Name In Schema` is correctly set.
 
 ### Version 1.0.3